adds remaining compatibility-methods for wx-versions<=3
